확장블록화 진행중
