+++ extern/util/static.js
@@ -380,10 +380,6 @@ EntryStatic.getAllBlocks = function() {
             blocks: ['ai_noti_img', 'ai_noti_text'],
         },
         {
-            category: 'artificial_intelligence',
-            blocks: ['audio_title', 'check_microphone', 'speech_to_text', 'get_microphone_volume'],
-        },
-        {
             category: 'expansion',
             blocks: [
                 'expansionBlockAddButton',
@@ -793,7 +789,7 @@ EntryStatic.colorSet = {
             VARIABLE: '#CE38CE',
             HARDWARE: '#FFFFFF',
             EXPANSION: '#FF8888',
-            ARTIFICIAL_INTELLIGENCE: '#FF8888',
+            AI_UTILIZE: '#FF8888',
             HIDDEN: '#FFFFFF',
         },
     },
@@ -812,7 +808,7 @@ EntryStatic.colorSet = {
             FUNC: '#de5c04',
             HARDWARE: '#00b6b1',
             EXPANSION: '#ef6d6d',
-            ARTIFICIAL_INTELLIGENCE: '#8222ff',
+            AI_UTILIZE: '#8222ff',
             HIDDEN: '#8aa3b2',
         },
         lighten: {
@@ -829,7 +825,7 @@ EntryStatic.colorSet = {
             FUNC: '#ff7b22',
             HARDWARE: '#78d5d3',
             EXPANSION: '#ffaeae',
-            ARTIFICIAL_INTELLIGENCE: '#ffaeae',
+            AI_UTILIZE: '#ffaeae',
             HIDDEN: '#ffaeae',
         },
         darken: {
@@ -846,7 +842,7 @@ EntryStatic.colorSet = {
             FUNC: '#a14100',
             HARDWARE: '#008380',
             EXPANSION: '#c63f3f',
-            ARTIFICIAL_INTELLIGENCE: '#670bdd',
+            AI_UTILIZE: '#670bdd',
             HIDDEN: '#728997',
         },
         emphasize: {
