+++ src/css/components/block_menu.less
@@ -360,13 +360,13 @@
     color: #fff;
 }
 
-#entryCategoryartificial_intelligence {
-    .backgroundImage("artificial_intelligence.svg");
+#entryCategoryai_utilize {
+    .backgroundImage("ai_utilize.svg");
     background-repeat: no-repeat;
 }
 
-.entrySelectedCategory#entryCategoryartificial_intelligence {
-    .backgroundImage("artificial_intelligence_on.svg");
+.entrySelectedCategory#entryCategoryai_utilize {
+    .backgroundImage("ai_utilize.svg");
     background-color: #8222ff;
     border-color: #670bdd;
     color: #fff;
