+++ src/playground/blocks/index.js
@@ -21,6 +21,12 @@ const basicBlockList = [
     require('./block_ai'),
 ];
 
+Entry.AI_UTILIZE_BLOCK = {};
+require('./block_ai_utilize_audio');
+Entry.AI_UTILIZE_BLOCK_LIST = {
+    audio: Entry.AI_UTILIZE_BLOCK.audio,
+};
+
 Entry.EXPANSION_BLOCK = {};
 require('./block_expansion_weather');
 require('./block_expansion_festival');
@@ -67,9 +73,9 @@ module.exports = {
     getBlocks() {
         const hardwareModules = hardware.getHardwareModuleList();
         registerHardwareBlockToStatic(hardwareModules);
-        const basicAndExpansionBlockObjectList = getBlockObject(
-            basicBlockList.concat(Object.values(Entry.EXPANSION_BLOCK_LIST))
-        );
+        basicBlockList.concat(Object.values(Entry.EXPANSION_BLOCK_LIST)),
+            basicBlockList.concat(Object.values(Entry.AI_UTILIZE_BLOCK_LIST));
+        const basicAndExpansionBlockObjectList = getBlockObject(basicBlockList);
         const hardwareBlockObjectList = getBlockObject(hardwareModules);
         return Object.assign({}, basicAndExpansionBlockObjectList, hardwareBlockObjectList);
     },
